refactor: Abolish hook extension methods, fix GroupBy inner-join filter, split helper tests, add MappingGenerate helper test (#320)
* Initial plan

* Fix PR reviews, abolish extension hooks, split test files, add HelperWithMappingGenerate test

Agent-Logs-Url: https://github.com/arika0093/Linqraft/sessions/0c81d629-367b-43f4-b367-dc53fab35297

---------
