+++ src/Linqraft.Core/SourceGenerator/ProjectionModelFinalizer.cs
@@ -193,6 +193,13 @@ internal static class ProjectionModelFinalizer
                     replacements,
                     cancellationToken
                 ),
+                InnerJoinFilterBodyText = requestTemplate.InnerJoinFilterBodyTemplate is null
+                    ? null
+                    : ProjectionBodyEmitter.ReplaceTokens(
+                        requestTemplate.InnerJoinFilterBodyTemplate,
+                        replacements,
+                        cancellationToken
+                    ),
             },
             GeneratedDtos = dtoReplacements.Select(replacement => replacement.Dto).ToArray(),
         };
