+++ src/Linqraft.Core/SourceGenerator/SourceWriters.cs
@@ -325,7 +325,10 @@ internal static class SourceWriters
                             ? expressionFieldName
                             : lambda;
                     var querySource = $"(({matchedQueryType})(object)query)";
-                    if (request.InnerJoinFilterBodyText is { } innerJoinFilterBodyText)
+                    if (
+                        request.InnerJoinFilterBodyText is { } innerJoinFilterBodyText
+                        && request.OperationKind != ProjectionOperationKind.GroupBy
+                    )
                     {
                         querySource =
                             $"{querySource}.Where({ProjectionBodyEmitter.AppendValueInline($"{request.SelectorParameterName} => ", innerJoinFilterBodyText)})";
@@ -510,16 +513,22 @@ internal static class SourceWriters
                     && request.ReceiverKind == ReceiverKind.IQueryable
                 )
                 {
+                    var querySource = request.InnerJoinFilterBodyText is { } innerJoinFilterBodyText
+                        ? $"source.Where({ProjectionBodyEmitter.AppendValueInline($"{request.SelectorParameterName} => ", innerJoinFilterBodyText)})"
+                        : "source";
                     builder.AppendLine(
-                        $"return source.Select({expressionFieldName});",
+                        $"return {querySource}.Select({expressionFieldName});",
                         cancellationToken
                     );
                 }
                 else
                 {
+                    var querySource = request.InnerJoinFilterBodyText is { } innerJoinFilterBodyText
+                        ? $"source.Where({ProjectionBodyEmitter.AppendValueInline($"{request.SelectorParameterName} => ", innerJoinFilterBodyText)})"
+                        : "source";
                     AppendMultilineLine(
                         builder,
-                        $"return source.Select({lambda});",
+                        $"return {querySource}.Select({lambda});",
                         cancellationToken
                     );
                 }
